Only warn about invalid DIA if coming from an active insulin plugin.
Since the id of the insulin plugin used is saved with each treatment record,
calculating IOB will ask such a plugin. If another insulin plugin is active now,
this would still raise an error about a too short DIA, which is confusing at best.
This change only displays this warning if the plugin is enabled.