Remove EventNewBg.isNew().
If a local BG source produces a new reading that is sent to NS, the
same reading is retrieved from NS and might be processed before
the local value is. That leads to the NS value creating a new
record, but since it's not the active BG source it doesn't trigger
a loop run. When the local value is processed it doesn't trigger a
loop run either since it wasn't a value.

This commit therefore removes the isNew field. Consumers query
isFromActiveBgSource() and isCurrent() as neede, which should be
enough.

Fixes #782, #783.